Build the swimlane stream comparison into the browser
The feature moves from the ruby_event_store-browser-swimlane contrib
gem into ruby_event_store-browser itself, available out of the box
with no configuration. The /swimlane and /swimlane/more routes are
served by App, the views live under views/swimlane, the url builders
on Urls, the stimulus controllers ship in the main js bundle and the
stream page always offers a Streamline entry link, followed by any
extension-contributed links.

The extension API stays intact - only its swimlane consumer is gone,
together with the contrib gem and its CI workflows. Specs follow into
the browser suite; new Urls and App subjects are mutation-covered.
