--- .github/workflows/ruby_event_store-browser-swimlane_coverage.yml
@@ -1,54 +1,0 @@
-name: ruby_event_store-browser-swimlane_coverage
-on:
-  workflow_dispatch:
-  repository_dispatch:
-    types:
-    - script
-  push:
-    branches:
-    - master
-    paths:
-    - contrib/ruby_event_store-browser-swimlane/Gemfile.lock
-    - ".github/workflows/ruby_event_store-browser-swimlane_coverage.yml"
-    - support/**
-    - "!support/bundler/**"
-    - "!support/ci/**"
-  pull_request:
-    paths:
-    - contrib/ruby_event_store-browser-swimlane/Gemfile.lock
-    - ".github/workflows/ruby_event_store-browser-swimlane_coverage.yml"
-    - support/**
-    - "!support/bundler/**"
-    - "!support/ci/**"
-  schedule:
-  - cron: 0 17 * * *
-jobs:
-  coverage:
-    runs-on: macos-14
-    timeout-minutes: 120
-    env:
-      WORKING_DIRECTORY: contrib/ruby_event_store-browser-swimlane
-      RUBY_VERSION: "${{ matrix.ruby_version }}"
-      BUNDLE_GEMFILE: "${{ matrix.bundle_gemfile }}"
-      BUNDLE_WITHOUT: database
-    strategy:
-      fail-fast: false
-      matrix:
-        include:
-        - ruby_version: ruby-4.0
-          bundle_gemfile: Gemfile
-    steps:
-    - uses: actions/checkout@v4
-      with:
-        fetch-depth: 1
-    - run: test -e ${{ env.BUNDLE_GEMFILE }}.lock
-      working-directory: "${{ env.WORKING_DIRECTORY }}"
-    - uses: ruby/setup-ruby@v1
-      with:
-        ruby-version: "${{ env.RUBY_VERSION }}"
-        bundler-cache: true
-        working-directory: "${{ env.WORKING_DIRECTORY }}"
-    - run: make mutate
-      working-directory: "${{ env.WORKING_DIRECTORY }}"
-      env:
-        RUBYOPT: "--enable-frozen-string-literal"
