--- .github/workflows/ruby_event_store-browser-swimlane_test.yml
@@ -1,55 +1,0 @@
-name: ruby_event_store-browser-swimlane_test
-on:
-  workflow_dispatch:
-  repository_dispatch:
-    types:
-    - script
-  push:
-    branches:
-    - master
-    paths:
-    - contrib/ruby_event_store-browser-swimlane/**
-    - ".github/workflows/ruby_event_store-browser-swimlane_test.yml"
-    - support/**
-    - "!support/bundler/**"
-    - "!support/ci/**"
-  pull_request:
-    paths:
-    - contrib/ruby_event_store-browser-swimlane/**
-    - ".github/workflows/ruby_event_store-browser-swimlane_test.yml"
-    - support/**
-    - "!support/bundler/**"
-    - "!support/ci/**"
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    timeout-minutes: 120
-    env:
-      WORKING_DIRECTORY: contrib/ruby_event_store-browser-swimlane
-      RUBY_VERSION: "${{ matrix.ruby_version }}"
-      BUNDLE_GEMFILE: "${{ matrix.bundle_gemfile }}"
-    strategy:
-      fail-fast: false
-      matrix:
-        include:
-        - ruby_version: ruby-4.0
-          bundle_gemfile: Gemfile
-        - ruby_version: ruby-3.4
-          bundle_gemfile: Gemfile
-        - ruby_version: ruby-3.3
-          bundle_gemfile: Gemfile
-    steps:
-    - uses: actions/checkout@v4
-      with:
-        fetch-depth: 1
-    - run: test -e ${{ env.BUNDLE_GEMFILE }}.lock
-      working-directory: "${{ env.WORKING_DIRECTORY }}"
-    - uses: ruby/setup-ruby@v1
-      with:
-        ruby-version: "${{ env.RUBY_VERSION }}"
-        bundler-cache: true
-        working-directory: "${{ env.WORKING_DIRECTORY }}"
-    - run: make test
-      working-directory: "${{ env.WORKING_DIRECTORY }}"
-      env:
-        RUBYOPT: "--enable-frozen-string-literal"
