--- contrib/ruby_event_store-browser-swimlane/.mutant.yml
@@ -1,17 +1,0 @@
-# https://github.com/mbj/mutant/blob/master/docs/configuration.md
-
-usage: opensource
-requires:
-  - ruby_event_store/browser/app
-  - ruby_event_store/browser/swimlane
-includes:
-  - lib
-integration:
-  name: rspec
-mutation:
-  operators: light
-coverage_criteria:
-  process_abort: true
-matcher:
-  subjects:
-    - RubyEventStore::Browser::Swimlane*
