--- contrib/ruby_event_store-browser-swimlane/Gemfile
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-source "https://rubygems.org"
-gemspec
-
-eval_gemfile "../../support/bundler/Gemfile.shared"
-gem "ruby_event_store", path: "../.."
-gem "ruby_event_store-browser", path: "../.."
