--- contrib/ruby_event_store-browser-swimlane/lib/ruby_event_store/browser/swimlane/version.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module RubyEventStore
-  module Browser
-    class Swimlane
-      VERSION = "0.1.0"
-    end
-  end
-end
