--- contrib/ruby_event_store-browser-swimlane/spec/spec_helper.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "../../../support/helpers/rspec_defaults"
-require "ruby_event_store"
-require "ruby_event_store/browser/app"
-require "ruby_event_store/browser/swimlane"
-require "rack"
-
-DummyEvent = Class.new(::RubyEventStore::Event)
