+++ ruby_event_store-browser/lib/ruby_event_store/browser/views/swimlane/_rows.html.erb
