+++ ruby_event_store-browser/spec/urls_spec.rb
@@ -90,6 +90,34 @@ module RubyEventStore
         )
       end
 
+      specify "swimlane_url" do
+        urls = Urls.from_configuration("http://example.com:9393", "/res")
+        expect(urls.swimlane_url(%w[fizz buzz])).to eq(
+          "http://example.com:9393/res/swimlane?streams%5B%5D=fizz&streams%5B%5D=buzz",
+        )
+      end
+
+      specify "swimlane_url with sort" do
+        urls = Urls.from_configuration("http://example.com:9393", nil)
+        expect(urls.swimlane_url(%w[fizz], "as_of")).to eq(
+          "http://example.com:9393/swimlane?streams%5B%5D=fizz&sort=as_of",
+        )
+      end
+
+      specify "swimlane_more_url" do
+        urls = Urls.from_configuration("http://example.com:9393", nil)
+        expect(urls.swimlane_more_url(%w[fizz buzz], "2024-01-01T12:00:00.000000Z", nil)).to eq(
+          "http://example.com:9393/swimlane/more?streams%5B%5D=fizz&streams%5B%5D=buzz&cursor=2024-01-01T12%3A00%3A00.000000Z",
+        )
+      end
+
+      specify "swimlane_more_url with sort" do
+        urls = Urls.from_configuration("http://example.com:9393", nil)
+        expect(urls.swimlane_more_url(%w[fizz], "2024-01-01T12:00:00.000000Z", "as_of")).to eq(
+          "http://example.com:9393/swimlane/more?streams%5B%5D=fizz&cursor=2024-01-01T12%3A00%3A00.000000Z&sort=as_of",
+        )
+      end
+
       specify "browser_css_url" do
         urls = Urls.from_configuration("http://example.com:9393", "/res")
         expect(urls.browser_css_url).to eq("http://example.com:9393/res/#{BROWSER_CSS}")
